Refactor API requests and remove embedded key
