+++ config.env
@@ -1,2 +1,2 @@
 # Copy this file to config.env and replace YOUR_API_KEY with your actual key
-COINGLASS_API_KEY="53b0a3236d8d4d2b9fff517c70c544ea"
+COINGLASS_API_KEY="YOUR_API_KEY"
