+++ current_oi.py
@@ -1,24 +1,10 @@
 import argparse
-import requests
 import csv
-import json
 import os
-import urllib.parse
-import urllib.request
-import requests
 
-BASE_URL = "https://open-api-v4.coinglass.com"
+from api_utils import fetch
 
 
-def fetch(endpoint: str, params=None, api_key=None) -> dict:
-    url = urllib.parse.urljoin(BASE_URL, endpoint)
-    headers = {}
-    if api_key:
-        headers["CG-API-KEY"] = api_key
-    resp = requests.get(url, headers=headers, params=params)
-    if resp.status_code != 200:
-        raise RuntimeError(f"Request failed: {resp.status_code}")
-    return resp.json()
 
 
 def save_dict(data: dict, filepath: str) -> None:
